backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
clerk-backend-api==4.2.0
cryptography==45.0.7
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
gunicorn
